Move p2pds back under p2pclient
Move back to tests/libp2p/p2pclient/test_p2pclient_integration.py
